use env variables for database & migations path
