+++ apps/dashboard/.env.development
@@ -1,0 +1,1 @@
+DATA_PATH=../../data
