+++ apps/dashboard/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@astrojs/node": "^9.2.2",
     "@teamgalena/shared": "workspace:*",
+    "@teamgalena/configs": "workspace:*",
     "arctic": "^3.7.0",
     "astro": "^5.10.1",
     "jsonwebtoken": "^9.0.2"
