+++ apps/dashboard/src/lib/server/discord.ts
@@ -1,6 +1,6 @@
+import { requireEnv } from "@teamgalena/shared/config";
 import { Discord, generateState } from "arctic";
 import type { APIContext } from "astro";
-import { requireEnv } from "./env";
 
 // TODO: Update redirect URI
 const discord = new Discord(
