+++ packages/configs/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@teamgalena/configs",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./tsconfig": "./tsconfig.json",
+    "./eslint": "./eslint.config.mjs"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.28.0",
+    "eslint": "^9.28.0",
+    "typescript-eslint": "^8.33.0"
+  },
+  "dependencies": {
+    "sqlite": "^5.1.1",
+    "sqlite3": "^5.1.7",
+    "chalk": "^5.3.0"
+  }
+}
