+++ packages/seeding/.env.development
@@ -1,0 +1,2 @@
+MIGRATIONS_PATH=../../migrations
+DATA_PATH=../../data
