+++ packages/seeding/tsconfig.json
@@ -1,0 +1,6 @@
+{
+   "extends": "@teamgalena/configs/tsconfig",
+   "include": [
+      "src"
+   ]
+}
