+++ packages/shared/eslint.config.mjs
@@ -1,16 +1,1 @@
-// @ts-check
-
-import eslint from "@eslint/js";
-import tseslint from "typescript-eslint";
-
-export default tseslint.config(
-  eslint.configs.recommended,
-  tseslint.configs.strict,
-  tseslint.configs.stylistic,
-  {
-    rules: {
-      "@typescript-eslint/consistent-type-definitions": ["error", "type"],
-      "no-console": "error",
-    },
-  }
-);
+export { default } from "@teamgalena/configs/eslint";
