+++ packages/shared/src/config.ts
@@ -1,7 +1,9 @@
+class ConfigurationError extends Error {}
+
 export function requireEnv(key: string) {
   const value = import.meta.env[key];
   if (value) return value;
-  throw new Error(`Missing environment variable '${key}'`);
+  throw new ConfigurationError(`Missing environment variable '${key}'`);
 }
 
 export function optionalEnv(key: string) {
